отдельный слайс для лидеров
