+++ packages/client/src/store/slices/index.ts
@@ -1,5 +1,7 @@
 import { authReducer } from './auth'
+import { leaderReducer } from './leader'
 
 export const rootReducer = {
   auth: authReducer,
+  leader: leaderReducer,
 }
