+++ packages/client/src/store/slices/leader/interfaces.ts
@@ -1,0 +1,7 @@
+export interface Leader {
+  data: {
+    name: string
+    avatar: string
+    score: string
+  }
+}
